Move mysql related classes from scaling-mysql to cdc-mysql
