+++ shardingsphere-kernel/shardingsphere-schedule/shardingsphere-cdc/shardingsphere-cdc-dialect/shardingsphere-cdc-mysql/src/main/java/org/apache/shardingsphere/cdc/mysql/MySQLInventoryDumper.java
