+++ shardingsphere-kernel/shardingsphere-schedule/shardingsphere-cdc/shardingsphere-cdc-dialect/shardingsphere-cdc-mysql/src/main/java/org/apache/shardingsphere/cdc/mysql/MySQLPositionInitializer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.mysql.component;
+package org.apache.shardingsphere.cdc.mysql;
 
 import com.google.common.base.Preconditions;
 import org.apache.shardingsphere.cdc.core.position.PositionInitializer;
